refactor: move checks for newer semversions to utilities
